requirements.txt
facenet-pytorch
torch
torchvision
opencv-python
numpy
pillow
scikit-learn
joblib
tqdm
